+++ src/components/NavMenu/NavMenu.js
@@ -1,14 +1,11 @@
-import { useState } from 'react';
-import logo from "../img/logo.png";
-import * as S from './styled/NavMenu.Style';
-import { Link } from 'react-router-dom';
-
+import { useState } from "react";
+import logo from "../../img/logo.png";
+import * as S from "./NavMenu.Style";
+import { Link } from "react-router-dom";
 
 export function NavMenu() {
-
   const [showMore, setShowMore] = useState(false);
 
-  
   function handleMoreClick() {
     setShowMore(!showMore);
   }
@@ -16,8 +13,8 @@ export function NavMenu() {
   return (
     <S.MainNav>
       <S.NavLogo>
-      <Link to="/">
-        <S.NavImage src={logo} alt="logo" />
+        <Link to="/">
+          <S.NavImage src={logo} alt="logo" />
         </Link>
       </S.NavLogo>
       <S.NavBurger onClick={handleMoreClick}>
@@ -26,24 +23,18 @@ export function NavMenu() {
         <S.BurgerLine></S.BurgerLine>
       </S.NavBurger>
       <S.NavMenu>
-      <S.MenuList showMore={showMore}>  
+        <S.MenuList showMore={showMore}>
           <S.NavMenuItem>
-            <S.NavMenuLink to="/">
-              Главное
-            </S.NavMenuLink>
+            <S.NavMenuLink to="/">Главное</S.NavMenuLink>
           </S.NavMenuItem>
           <S.NavMenuItem>
-            <S.NavMenuLink to="/favorites">
-              Мой плейлист
-            </S.NavMenuLink>
+            <S.NavMenuLink to="/favorites">Мой плейлист</S.NavMenuLink>
           </S.NavMenuItem>
           <S.NavMenuItem>
-          <S.NavMenuLink to="/signin">
-              Войти
-            </S.NavMenuLink>
+            <S.NavMenuLink to="/signin">Войти</S.NavMenuLink>
           </S.NavMenuItem>
         </S.MenuList>
-      </S.NavMenu> 
+      </S.NavMenu>
     </S.MainNav>
   );
 }
