+++ src/components/center/Center.style.js
