+++ src/components/player/player.style.js
