+++ src/components/righPanel/rightpanel.style.js
